fixed imports

+++ pysapi/tools/structures/ovh.py
@@ -1,4 +1,3 @@
-from .. import Double
 import numpy as np
 
 def cumovh(structure_set, ptv_id, oar_id, min_mm=-10.0, max_mm=50.0, step_mm=5.0,norm=True):
@@ -22,7 +21,7 @@ def cumovh(structure_set, ptv_id, oar_id, min_mm=-10.0, max_mm=50.0, step_mm=5.0
 
     try:
         overlap = structure_set.AddStructure('DOSE_REGION','temp_pysapi2')
-        ovh_mm = np.arange(min_mm, max_mm, step_mm, dtype=Double)
+        ovh_mm = np.arange(min_mm, max_mm, step_mm)
         ovh_vol = []
 
         for expanded_mm in ovh_mm:
